Convert livedata_server in @ddp-server to Class structure w/ minimal performance improvements
