+++ packages/ddp-server/livedata_server.js
@@ -1,800 +1,21 @@
-import Session from "./session";
+import ServerClass from "./server";
 import SessionCollectionView from "./session_collection_view";
 import SessionDocumentView from "./session_document_view";
 
 DDPServer = {};
 
-var Fiber = Npm.require('fibers');
-
-// This file contains classes:
+// This file links these classes:
 // * Session - The server's connection to a single DDP client
 // * Subscription - A single subscription for a single client
 // * Server - An entire server that may talk to > 1 client. A DDP endpoint.
 //
-// Session and Subscription are file scope. For now, until we freeze
-// the interface, Server is package scope (in the future it should be
-// exported.)
 
-// Represents a single document in a SessionCollectionView
 DDPServer._SessionDocumentView = SessionDocumentView;
-
-/**
- * Represents a client's view of a single collection
- * @param {String} collectionName Name of the collection it represents
- * @param {Object.<String, Function>} sessionCallbacks The callbacks for added, changed, removed
- * @class SessionCollectionView
- */
 
 DDPServer._SessionCollectionView = SessionCollectionView;
 
-/******************************************************************************/
-/* Subscription                                                               */
-/******************************************************************************/
 
-// ctor for a sub handle: the input to each publish function
-
-// Instance name is this because it's usually referred to as this inside a
-// publish
-/**
- * @summary The server's side of a subscription
- * @class Subscription
- * @instanceName this
- * @showInstanceName true
- */
-var Subscription = function (
-    session, handler, subscriptionId, params, name) {
-  var self = this;
-  self._session = session; // type is Session
-
-  /**
-   * @summary Access inside the publish function. The incoming [connection](#meteor_onconnection) for this subscription.
-   * @locus Server
-   * @name  connection
-   * @memberOf Subscription
-   * @instance
-   */
-  self.connection = session.connectionHandle; // public API object
-
-  self._handler = handler;
-
-  // my subscription ID (generated by client, undefined for universal subs).
-  self._subscriptionId = subscriptionId;
-  // undefined for universal subs
-  self._name = name;
-
-  self._params = params || [];
-
-  // Only named subscriptions have IDs, but we need some sort of string
-  // internally to keep track of all subscriptions inside
-  // SessionDocumentViews. We use this subscriptionHandle for that.
-  if (self._subscriptionId) {
-    self._subscriptionHandle = 'N' + self._subscriptionId;
-  } else {
-    self._subscriptionHandle = 'U' + Random.id();
-  }
-
-  // has _deactivate been called?
-  self._deactivated = false;
-
-  // stop callbacks to g/c this sub.  called w/ zero arguments.
-  self._stopCallbacks = [];
-
-  // the set of (collection, documentid) that this subscription has
-  // an opinion about
-  self._documents = new Map();
-
-  // remember if we are ready.
-  self._ready = false;
-
-  // Part of the public API: the user of this sub.
-
-  /**
-   * @summary Access inside the publish function. The id of the logged-in user, or `null` if no user is logged in.
-   * @locus Server
-   * @memberOf Subscription
-   * @name  userId
-   * @instance
-   */
-  self.userId = session.userId;
-
-  // For now, the id filter is going to default to
-  // the to/from DDP methods on MongoID, to
-  // specifically deal with mongo/minimongo ObjectIds.
-
-  // Later, you will be able to make this be "raw"
-  // if you want to publish a collection that you know
-  // just has strings for keys and no funny business, to
-  // a ddp consumer that isn't minimongo
-
-  self._idFilter = {
-    idStringify: MongoID.idStringify,
-    idParse: MongoID.idParse
-  };
-
-  Package['facts-base'] && Package['facts-base'].Facts.incrementServerFact(
-    "livedata", "subscriptions", 1);
-};
-
-Object.assign(Subscription.prototype, {
-  _runHandler: function () {
-    // XXX should we unblock() here? Either before running the publish
-    // function, or before running _publishCursor.
-    //
-    // Right now, each publish function blocks all future publishes and
-    // methods waiting on data from Mongo (or whatever else the function
-    // blocks on). This probably slows page load in common cases.
-
-    var self = this;
-    try {
-      var res = DDP._CurrentPublicationInvocation.withValue(
-        self,
-        () => maybeAuditArgumentChecks(
-          self._handler, self, EJSON.clone(self._params),
-          // It's OK that this would look weird for universal subscriptions,
-          // because they have no arguments so there can never be an
-          // audit-argument-checks failure.
-          "publisher '" + self._name + "'"
-        )
-      );
-    } catch (e) {
-      self.error(e);
-      return;
-    }
-
-    // Did the handler call this.error or this.stop?
-    if (self._isDeactivated())
-      return;
-
-    self._publishHandlerResult(res);
-  },
-
-  _publishHandlerResult: function (res) {
-    // SPECIAL CASE: Instead of writing their own callbacks that invoke
-    // this.added/changed/ready/etc, the user can just return a collection
-    // cursor or array of cursors from the publish function; we call their
-    // _publishCursor method which starts observing the cursor and publishes the
-    // results. Note that _publishCursor does NOT call ready().
-    //
-    // XXX This uses an undocumented interface which only the Mongo cursor
-    // interface publishes. Should we make this interface public and encourage
-    // users to implement it themselves? Arguably, it's unnecessary; users can
-    // already write their own functions like
-    //   var publishMyReactiveThingy = function (name, handler) {
-    //     Meteor.publish(name, function () {
-    //       var reactiveThingy = handler();
-    //       reactiveThingy.publishMe();
-    //     });
-    //   };
-
-    var self = this;
-    var isCursor = function (c) {
-      return c && c._publishCursor;
-    };
-    if (isCursor(res)) {
-      try {
-        res._publishCursor(self);
-      } catch (e) {
-        self.error(e);
-        return;
-      }
-      // _publishCursor only returns after the initial added callbacks have run.
-      // mark subscription as ready.
-      self.ready();
-    } else if (Array.isArray(res)) {
-      // check all the elements are cursors
-      if (! _.all(res, isCursor)) {
-        self.error(new Error("Publish function returned an array of non-Cursors"));
-        return;
-      }
-      // find duplicate collection names
-      // XXX we should support overlapping cursors, but that would require the
-      // merge box to allow overlap within a subscription
-      var collectionNames = {};
-      for (var i = 0; i < res.length; ++i) {
-        var collectionName = res[i]._getCollectionName();
-        if (_.has(collectionNames, collectionName)) {
-          self.error(new Error(
-            "Publish function returned multiple cursors for collection " +
-              collectionName));
-          return;
-        }
-        collectionNames[collectionName] = true;
-      };
-
-      try {
-      res.forEach(function (cur) {
-          cur._publishCursor(self);
-        });
-      } catch (e) {
-        self.error(e);
-        return;
-      }
-      self.ready();
-    } else if (res) {
-      // truthy values other than cursors or arrays are probably a
-      // user mistake (possible returning a Mongo document via, say,
-      // `coll.findOne()`).
-      self.error(new Error("Publish function can only return a Cursor or "
-                           + "an array of Cursors"));
-    }
-  },
-
-  // This calls all stop callbacks and prevents the handler from updating any
-  // SessionCollectionViews further. It's used when the user unsubscribes or
-  // disconnects, as well as during setUserId re-runs. It does *NOT* send
-  // removed messages for the published objects; if that is necessary, call
-  // _removeAllDocuments first.
-  _deactivate: function() {
-    var self = this;
-    if (self._deactivated)
-      return;
-    self._deactivated = true;
-    self._callStopCallbacks();
-    Package['facts-base'] && Package['facts-base'].Facts.incrementServerFact(
-      "livedata", "subscriptions", -1);
-  },
-
-  _callStopCallbacks: function () {
-    var self = this;
-    // tell listeners, so they can clean up
-    var callbacks = self._stopCallbacks;
-    self._stopCallbacks = [];
-    callbacks.forEach(function (callback) {
-      callback();
-    });
-  },
-
-  // Send remove messages for every document.
-  _removeAllDocuments: function () {
-    var self = this;
-    Meteor._noYieldsAllowed(function () {
-      self._documents.forEach(function (collectionDocs, collectionName) {
-        collectionDocs.forEach(function (strId) {
-          self.removed(collectionName, self._idFilter.idParse(strId));
-        });
-      });
-    });
-  },
-
-  // Returns a new Subscription for the same session with the same
-  // initial creation parameters. This isn't a clone: it doesn't have
-  // the same _documents cache, stopped state or callbacks; may have a
-  // different _subscriptionHandle, and gets its userId from the
-  // session, not from this object.
-  _recreate: function () {
-    var self = this;
-    return new Subscription(
-      self._session, self._handler, self._subscriptionId, self._params,
-      self._name);
-  },
-
-  /**
-   * @summary Call inside the publish function.  Stops this client's subscription, triggering a call on the client to the `onStop` callback passed to [`Meteor.subscribe`](#meteor_subscribe), if any. If `error` is not a [`Meteor.Error`](#meteor_error), it will be [sanitized](#meteor_error).
-   * @locus Server
-   * @param {Error} error The error to pass to the client.
-   * @instance
-   * @memberOf Subscription
-   */
-  error: function (error) {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    self._session._stopSubscription(self._subscriptionId, error);
-  },
-
-  // Note that while our DDP client will notice that you've called stop() on the
-  // server (and clean up its _subscriptions table) we don't actually provide a
-  // mechanism for an app to notice this (the subscribe onError callback only
-  // triggers if there is an error).
-
-  /**
-   * @summary Call inside the publish function.  Stops this client's subscription and invokes the client's `onStop` callback with no error.
-   * @locus Server
-   * @instance
-   * @memberOf Subscription
-   */
-  stop: function () {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    self._session._stopSubscription(self._subscriptionId);
-  },
-
-  /**
-   * @summary Call inside the publish function.  Registers a callback function to run when the subscription is stopped.
-   * @locus Server
-   * @memberOf Subscription
-   * @instance
-   * @param {Function} func The callback function
-   */
-  onStop: function (callback) {
-    var self = this;
-    callback = Meteor.bindEnvironment(callback, 'onStop callback', self);
-    if (self._isDeactivated())
-      callback();
-    else
-      self._stopCallbacks.push(callback);
-  },
-
-  // This returns true if the sub has been deactivated, *OR* if the session was
-  // destroyed but the deferred call to _deactivateAllSubscriptions hasn't
-  // happened yet.
-  _isDeactivated: function () {
-    var self = this;
-    return self._deactivated || self._session.inQueue === null;
-  },
-
-  /**
-   * @summary Call inside the publish function.  Informs the subscriber that a document has been added to the record set.
-   * @locus Server
-   * @memberOf Subscription
-   * @instance
-   * @param {String} collection The name of the collection that contains the new document.
-   * @param {String} id The new document's ID.
-   * @param {Object} fields The fields in the new document.  If `_id` is present it is ignored.
-   */
-  added: function (collectionName, id, fields) {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    id = self._idFilter.idStringify(id);
-    let ids = self._documents.get(collectionName);
-    if (ids == null) {
-      ids = new Set();
-      self._documents.set(collectionName, ids);
-    }
-    ids.add(id);
-    self._session.added(self._subscriptionHandle, collectionName, id, fields);
-  },
-
-  /**
-   * @summary Call inside the publish function.  Informs the subscriber that a document in the record set has been modified.
-   * @locus Server
-   * @memberOf Subscription
-   * @instance
-   * @param {String} collection The name of the collection that contains the changed document.
-   * @param {String} id The changed document's ID.
-   * @param {Object} fields The fields in the document that have changed, together with their new values.  If a field is not present in `fields` it was left unchanged; if it is present in `fields` and has a value of `undefined` it was removed from the document.  If `_id` is present it is ignored.
-   */
-  changed: function (collectionName, id, fields) {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    id = self._idFilter.idStringify(id);
-    self._session.changed(self._subscriptionHandle, collectionName, id, fields);
-  },
-
-  /**
-   * @summary Call inside the publish function.  Informs the subscriber that a document has been removed from the record set.
-   * @locus Server
-   * @memberOf Subscription
-   * @instance
-   * @param {String} collection The name of the collection that the document has been removed from.
-   * @param {String} id The ID of the document that has been removed.
-   */
-  removed: function (collectionName, id) {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    id = self._idFilter.idStringify(id);
-    // We don't bother to delete sets of things in a collection if the
-    // collection is empty.  It could break _removeAllDocuments.
-    self._documents.get(collectionName).delete(id);
-    self._session.removed(self._subscriptionHandle, collectionName, id);
-  },
-
-  /**
-   * @summary Call inside the publish function.  Informs the subscriber that an initial, complete snapshot of the record set has been sent.  This will trigger a call on the client to the `onReady` callback passed to  [`Meteor.subscribe`](#meteor_subscribe), if any.
-   * @locus Server
-   * @memberOf Subscription
-   * @instance
-   */
-  ready: function () {
-    var self = this;
-    if (self._isDeactivated())
-      return;
-    if (!self._subscriptionId)
-      return;  // unnecessary but ignored for universal sub
-    if (!self._ready) {
-      self._session.sendReady([self._subscriptionId]);
-      self._ready = true;
-    }
-  }
-});
-
-/******************************************************************************/
-/* Server                                                                     */
-/******************************************************************************/
-
-Server = function (options) {
-  var self = this;
-
-  // The default heartbeat interval is 30 seconds on the server and 35
-  // seconds on the client.  Since the client doesn't need to send a
-  // ping as long as it is receiving pings, this means that pings
-  // normally go from the server to the client.
-  //
-  // Note: Troposphere depends on the ability to mutate
-  // Meteor.server.options.heartbeatTimeout! This is a hack, but it's life.
-  self.options = _.defaults(options || {}, {
-    heartbeatInterval: 15000,
-    heartbeatTimeout: 15000,
-    // For testing, allow responding to pings to be disabled.
-    respondToPings: true
-  });
-
-  // Map of callbacks to call when a new connection comes in to the
-  // server and completes DDP version negotiation. Use an object instead
-  // of an array so we can safely remove one from the list while
-  // iterating over it.
-  self.onConnectionHook = new Hook({
-    debugPrintExceptions: "onConnection callback"
-  });
-
-  // Map of callbacks to call when a new message comes in.
-  self.onMessageHook = new Hook({
-    debugPrintExceptions: "onMessage callback"
-  });
-
-  self.publish_handlers = {};
-
-  self.method_handlers = {};
-  
-  self.universal_publish_handlers = [];
-
-  self.sessions = new Map(); // map from id to session
-
-  self.stream_server = new StreamServer;
-
-  self.stream_server.register(function (socket) {
-    // socket implements the SockJSConnection interface
-    socket._meteorSession = null;
-
-    var sendError = function (reason, offendingMessage) {
-      var msg = {msg: 'error', reason: reason};
-      if (offendingMessage)
-        msg.offendingMessage = offendingMessage;
-      socket.send(DDPCommon.stringifyDDP(msg));
-    };
-
-    socket.on('data', function (raw_msg) {
-      if (Meteor._printReceivedDDP) {
-        Meteor._debug("Received DDP", raw_msg);
-      }
-      try {
-        try {
-          var msg = DDPCommon.parseDDP(raw_msg);
-        } catch (err) {
-          sendError('Parse error');
-          return;
-        }
-        if (msg === null || !msg.msg) {
-          sendError('Bad request', msg);
-          return;
-        }
-
-        if (msg.msg === 'connect') {
-          if (socket._meteorSession) {
-            sendError("Already connected", msg);
-            return;
-          }
-          Fiber(function () {
-            self._handleConnect(socket, msg);
-          }).run();
-          return;
-        }
-
-        if (!socket._meteorSession) {
-          sendError('Must connect first', msg);
-          return;
-        }
-        socket._meteorSession.processMessage(msg);
-      } catch (e) {
-        // XXX print stack nicely
-        Meteor._debug("Internal exception while processing message", msg, e);
-      }
-    });
-
-    socket.on('close', function () {
-      if (socket._meteorSession) {
-        Fiber(function () {
-          socket._meteorSession.close();
-        }).run();
-      }
-    });
-  });
-};
-
-Object.assign(Server.prototype, {
-
-  /**
-   * @summary Register a callback to be called when a new DDP connection is made to the server.
-   * @locus Server
-   * @param {function} callback The function to call when a new DDP connection is established.
-   * @memberOf Meteor
-   * @importFromPackage meteor
-   */
-  onConnection: function (fn) {
-    var self = this;
-    return self.onConnectionHook.register(fn);
-  },
-
-  /**
-   * @summary Register a callback to be called when a new DDP message is received.
-   * @locus Server
-   * @param {function} callback The function to call when a new DDP message is received.
-   * @memberOf Meteor
-   * @importFromPackage meteor
-   */
-  onMessage: function (fn) {
-    var self = this;
-    return self.onMessageHook.register(fn);
-  },
-
-  _handleConnect: function (socket, msg) {
-    var self = this;
-
-    // The connect message must specify a version and an array of supported
-    // versions, and it must claim to support what it is proposing.
-    if (!(typeof (msg.version) === 'string' &&
-          Array.isArray(msg.support) &&
-          _.all(msg.support, _.isString) &&
-          _.contains(msg.support, msg.version))) {
-      socket.send(DDPCommon.stringifyDDP({msg: 'failed',
-                                version: DDPCommon.SUPPORTED_DDP_VERSIONS[0]}));
-      socket.close();
-      return;
-    }
-
-    // In the future, handle session resumption: something like:
-    //  socket._meteorSession = self.sessions[msg.session]
-    var version = calculateVersion(msg.support, DDPCommon.SUPPORTED_DDP_VERSIONS);
-
-    if (msg.version !== version) {
-      // The best version to use (according to the client's stated preferences)
-      // is not the one the client is trying to use. Inform them about the best
-      // version to use.
-      socket.send(DDPCommon.stringifyDDP({msg: 'failed', version: version}));
-      socket.close();
-      return;
-    }
-
-    // Yay, version matches! Create a new session.
-    // Note: Troposphere depends on the ability to mutate
-    // Meteor.server.options.heartbeatTimeout! This is a hack, but it's life.
-    socket._meteorSession = new Session(self, version, socket, self.options);
-    self.sessions.set(socket._meteorSession.id, socket._meteorSession);
-    self.onConnectionHook.each(function (callback) {
-      if (socket._meteorSession)
-        callback(socket._meteorSession.connectionHandle);
-      return true;
-    });
-  },
-  /**
-   * Register a publish handler function.
-   *
-   * @param name {String} identifier for query
-   * @param handler {Function} publish handler
-   * @param options {Object}
-   *
-   * Server will call handler function on each new subscription,
-   * either when receiving DDP sub message for a named subscription, or on
-   * DDP connect for a universal subscription.
-   *
-   * If name is null, this will be a subscription that is
-   * automatically established and permanently on for all connected
-   * client, instead of a subscription that can be turned on and off
-   * with subscribe().
-   *
-   * options to contain:
-   *  - (mostly internal) is_auto: true if generated automatically
-   *    from an autopublish hook. this is for cosmetic purposes only
-   *    (it lets us determine whether to print a warning suggesting
-   *    that you turn off autopublish.)
-   */
-
-  /**
-   * @summary Publish a record set.
-   * @memberOf Meteor
-   * @importFromPackage meteor
-   * @locus Server
-   * @param {String|Object} name If String, name of the record set.  If Object, publications Dictionary of publish functions by name.  If `null`, the set has no name, and the record set is automatically sent to all connected clients.
-   * @param {Function} func Function called on the server each time a client subscribes.  Inside the function, `this` is the publish handler object, described below.  If the client passed arguments to `subscribe`, the function is called with the same arguments.
-   */
-  publish: function (name, handler, options) {
-    var self = this;
-
-    if (! _.isObject(name)) {
-      options = options || {};
-
-      if (name && name in self.publish_handlers) {
-        Meteor._debug("Ignoring duplicate publish named '" + name + "'");
-        return;
-      }
-
-      if (Package.autopublish && !options.is_auto) {
-        // They have autopublish on, yet they're trying to manually
-        // picking stuff to publish. They probably should turn off
-        // autopublish. (This check isn't perfect -- if you create a
-        // publish before you turn on autopublish, it won't catch
-        // it. But this will definitely handle the simple case where
-        // you've added the autopublish package to your app, and are
-        // calling publish from your app code.)
-        if (!self.warned_about_autopublish) {
-          self.warned_about_autopublish = true;
-          Meteor._debug(
-    "** You've set up some data subscriptions with Meteor.publish(), but\n" +
-    "** you still have autopublish turned on. Because autopublish is still\n" +
-    "** on, your Meteor.publish() calls won't have much effect. All data\n" +
-    "** will still be sent to all clients.\n" +
-    "**\n" +
-    "** Turn off autopublish by removing the autopublish package:\n" +
-    "**\n" +
-    "**   $ meteor remove autopublish\n" +
-    "**\n" +
-    "** .. and make sure you have Meteor.publish() and Meteor.subscribe() calls\n" +
-    "** for each collection that you want clients to see.\n");
-        }
-      }
-
-      if (name)
-        self.publish_handlers[name] = handler;
-      else {
-        self.universal_publish_handlers.push(handler);
-        // Spin up the new publisher on any existing session too. Run each
-        // session's subscription in a new Fiber, so that there's no change for
-        // self.sessions to change while we're running this loop.
-        self.sessions.forEach(function (session) {
-          if (!session._dontStartNewUniversalSubs) {
-            Fiber(function() {
-              session._startSubscription(handler);
-            }).run();
-          }
-        });
-      }
-    }
-    else{
-      Object.entries(name).forEach(function(value, key) {
-        self.publish(key, value, {});
-      });
-    }
-  },
-
-  _removeSession: function (session) {
-    var self = this;
-    self.sessions.delete(session.id);
-  },
-
-  /**
-   * @summary Defines functions that can be invoked over the network by clients.
-   * @locus Anywhere
-   * @param {Object} methods Dictionary whose keys are method names and values are functions.
-   * @memberOf Meteor
-   * @importFromPackage meteor
-   */
-  methods: function (methods) {
-    var self = this;
-    Object.entries(methods).forEach(async function (func, name) {
-      //Validate func param is function
-      if (typeof func !== 'function')
-        throw new Error("Method '" + name + "' must be a function");
-
-      //Verify method name uniqueness
-      if (self.method_handlers[name])
-        throw new Error("A method named '" + name + "' is already defined");
-
-      self.method_handlers[name] = func;
-    });
-  },
-
-  call: function (name, ...args) {
-    if (args.length && typeof args[args.length - 1] === "function") {
-      // If it's a function, the last argument is the result callback, not
-      // a parameter to the remote method.
-      var callback = args.pop();
-    }
-
-    return this.apply(name, args, callback);
-  },
-
-  // A version of the call method that always returns a Promise.
-  callAsync: function (name, ...args) {
-    return this.applyAsync(name, args);
-  },
-
-  apply: function (name, args, options, callback) {
-    // We were passed 3 arguments. They may be either (name, args, options)
-    // or (name, args, callback)
-    if (! callback && typeof options === 'function') {
-      callback = options;
-      options = {};
-    } else {
-      options = options || {};
-    }
-
-    const promise = this.applyAsync(name, args, options);
-
-    // Return the result in whichever way the caller asked for it. Note that we
-    // do NOT block on the write fence in an analogous way to how the client
-    // blocks on the relevant data being visible, so you are NOT guaranteed that
-    // cursor observe callbacks have fired when your callback is invoked. (We
-    // can change this if there's a real use case.)
-    if (callback) {
-      promise.then(
-        result => callback(undefined, result),
-        exception => callback(exception)
-      );
-    } else {
-      return promise.await();
-    }
-  },
-
-  /**
-  * @param options {Optional Object}
-  */
-  applyAsync: function (name, args, options) {
-    // Run the handler
-    var handler = this.method_handlers[name];
-    if (! handler) {
-      return Promise.reject(
-        new Meteor.Error(404, `Method '${name}' not found`)
-      );
-    }
-
-    // If this is a method call from within another method or publish function,
-    // get the user state from the outer method or publish function, otherwise
-    // don't allow setUserId to be called
-    var userId = null;
-    var setUserId = function() {
-      throw new Error("Can't call setUserId on a server initiated method call");
-    };
-    var connection = null;
-    var currentMethodInvocation = DDP._CurrentMethodInvocation.get();
-    var currentPublicationInvocation = DDP._CurrentPublicationInvocation.get();
-    var randomSeed = null;
-    if (currentMethodInvocation) {
-      userId = currentMethodInvocation.userId;
-      setUserId = function(userId) {
-        currentMethodInvocation.setUserId(userId);
-      };
-      connection = currentMethodInvocation.connection;
-      randomSeed = DDPCommon.makeRpcSeed(currentMethodInvocation, name);
-    } else if (currentPublicationInvocation) {
-      userId = currentPublicationInvocation.userId;
-      setUserId = function(userId) {
-        currentPublicationInvocation._session._setUserId(userId);
-      };
-      connection = currentPublicationInvocation.connection;
-    }
-
-    var invocation = new DDPCommon.MethodInvocation({
-      isSimulation: false,
-      userId,
-      setUserId,
-      connection,
-      randomSeed
-    });
-
-    return new Promise(resolve => resolve(
-      DDP._CurrentMethodInvocation.withValue(
-        invocation,
-        () => maybeAuditArgumentChecks(
-          handler, invocation, EJSON.clone(args),
-          "internal call to '" + name + "'"
-        )
-      )
-    )).then(EJSON.clone);
-  },
-
-  _urlForSession: function (sessionId) {
-    var self = this;
-    var session = self.sessions.get(sessionId);
-    if (session)
-      return session._socketUrl;
-    else
-      return null;
-  }
-});
+Server = new ServerClass();
 
 var calculateVersion = function (clientSupportedVersions,
                                  serverSupportedVersions) {
@@ -809,56 +30,3 @@ var calculateVersion = function (clientSupportedVersions,
 
 DDPServer._calculateVersion = calculateVersion;
 
-
-// "blind" exceptions other than those that were deliberately thrown to signal
-// errors to the client
-var wrapInternalException = function (exception, context) {
-  if (!exception) return exception;
-
-  // To allow packages to throw errors intended for the client but not have to
-  // depend on the Meteor.Error class, `isClientSafe` can be set to true on any
-  // error before it is thrown.
-  if (exception.isClientSafe) {
-    if (!(exception instanceof Meteor.Error)) {
-      const originalMessage = exception.message;
-      exception = new Meteor.Error(exception.error, exception.reason, exception.details);
-      exception.message = originalMessage;
-    }
-    return exception;
-  }
-
-  // Tests can set the '_expectedByTest' flag on an exception so it won't go to
-  // the server log.
-  if (!exception._expectedByTest) {
-    Meteor._debug("Exception " + context, exception.stack);
-    if (exception.sanitizedError) {
-      Meteor._debug("Sanitized and reported to the client as:", exception.sanitizedError);
-      Meteor._debug();
-    }
-  }
-
-  // Did the error contain more details that could have been useful if caught in
-  // server code (or if thrown from non-client-originated code), but also
-  // provided a "sanitized" version with more context than 500 Internal server
-  // error? Use that.
-  if (exception.sanitizedError) {
-    if (exception.sanitizedError.isClientSafe)
-      return exception.sanitizedError;
-    Meteor._debug("Exception " + context + " provides a sanitizedError that " +
-                  "does not have isClientSafe property set; ignoring");
-  }
-
-  return new Meteor.Error(500, "Internal server error");
-};
-
-
-// Audit argument checks, if the audit-argument-checks package exists (it is a
-// weak dependency of this package).
-var maybeAuditArgumentChecks = function (f, context, args, description) {
-  args = args || [];
-  if (Package['audit-argument-checks']) {
-    return Match._failIfArgumentsAreNotAllChecked(
-      f, context, args, description);
-  }
-  return f.apply(context, args);
-};
