+++ packages/ddp-server/server_convenience.js
@@ -1,9 +1,11 @@
+import ServerClass from "./server";
+
 if (process.env.DDP_DEFAULT_CONNECTION_URL) {
   __meteor_runtime_config__.DDP_DEFAULT_CONNECTION_URL =
     process.env.DDP_DEFAULT_CONNECTION_URL;
 }
 
-Meteor.server = new Server;
+Meteor.server = new ServerClass();
 
 Meteor.refresh = function (notification) {
   DDPServer._InvalidationCrossbar.fire(notification);
@@ -13,7 +15,7 @@ Meteor.refresh = function (notification) {
 // be called directly on Meteor.
 ['publish', 'methods', 'call', 'apply', 'onConnection', 'onMessage'].forEach(
        function (name) {
-         Meteor[name] = Meteor.server[name].bind(Meteor.server);
+         Meteor[name] = () => Meteor.server[name]
        });
 
 // Meteor.server used to be called Meteor.default_server. Provide
